+++ app/client/src/pages/organization/index.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { Switch, useRouteMatch, useLocation } from "react-router-dom";
 import PageWrapper from "pages/common/PageWrapper";
 import Settings from "./settings";
-import Invite from "./invite";
 import DefaultOrgPage from "./defaultOrgPage";
 import AppRoute from "pages/common/AppRoute";
 export const Organization = () => {
@@ -16,12 +15,6 @@ export const Organization = () => {
           path={`${path}/:orgId/settings`}
           component={Settings}
           name={"Settings"}
-        />
-        <AppRoute
-          exact
-          path={`${path}/invite`}
-          component={Invite}
-          name={"Invite"}
         />
         <AppRoute component={DefaultOrgPage} name={"DefaultOrgPage"} />
       </Switch>
